pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "databricks-agent-toolkit"
version = "0.1.33"
description = "Unified toolkit for building production agents on Databricks - pre-wired integrations for LangGraph, LangChain, and custom agents"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}

authors = [
    {name = "Databricks Agent Toolkit Team"}
]

keywords = [
    "databricks",
    "agents",
    "llm",
    "langgraph",
    "langchain",
    "mlflow",
    "unity-catalog",
    "lakebase",
    "model-serving",
    "mcp",
    "agent-toolkit"
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

dependencies = [
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "packaging>=23.0",
    "psutil>=5.9.0",
    "jinja2>=3.1.0",  # For scaffold generation
]

[project.optional-dependencies]
databricks = [
    "databricks-sdk>=0.20.0",
    "databricks-sql-connector>=3.0.0",
    "mlflow[databricks]>=3.1.0",
]

agent-frameworks = [
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-core>=0.2.0",
]

optimization = [
    "dspy-ai>=2.4.0",
    "textgrad>=0.1.0",
    "ollama>=0.1.0",
]

mcp = [
    "mcp>=0.1.0",
]

a2a = [
    "a2a-sdk>=0.3.0",
]

telemetry = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation>=0.41b0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
]

web = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
]

documents = [
    "PyPDF2>=3.0.0",
    "python-docx>=0.8.11",
]

scaffolding = [
    "jinja2>=3.1.0",
]

all = [
    "databricks-agent-toolkit[databricks,agent-frameworks,optimization,mcp,telemetry,web,documents,scaffolding]",
]

dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "ipython>=8.12.0",
    "jupyter>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/databricks/agent-toolkit"
Documentation = "https://databricks-agent-toolkit.readthedocs.io"
Repository = "https://github.com/databricks/agent-toolkit"
Issues = "https://github.com/databricks/agent-toolkit/issues"

[tool.setuptools]
packages.find = {where = ["."], include = ["databricks_agent_toolkit*", "shared*", "evaluation*", "memory*", "mcp*", "optimization*", "telemetry*", "monitoring*"]}

[tool.setuptools.package-data]
"*" = ["py.typed", "*.jinja2", "*.yaml", "*.yml"]

[project.scripts]
databricks-agent-toolkit = "databricks_agent_toolkit.cli.main:main"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "0.1.12"

[tool.mypy]
python_version = "0.1.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
